+++ gateway/crates/flow-processor/src/flow_executor.rs
@@ -1,25 +1,20 @@
 use tokio;
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::mpsc;
 use tokio::time::Duration;
 use tokio::task::JoinHandle;
 use crate::types::*;
-use crate::errors::FlowProcessorError;
+use crate::flow_router::FlowProcessorRouter;
 use persistent_storage::init::PostgresRepo;
 use tracing;
 use std::collections::HashMap;
 use uuid::Uuid;
 use tokio_util::sync::CancellationToken;
 
-#[derive(Clone)]
-pub struct FlowProcessorState {
-    storage: PostgresRepo,
-}
-
 pub struct FlowProcessor {
     pub tx_receiver: mpsc::Receiver<(FlowProcessorMessage, OneshotFlowProcessorSender)>,
     pub flow_receiver: mpsc::Receiver<Uuid>,
     pub flow_sender: mpsc::Sender<Uuid>,
-    pub state: FlowProcessorState,
+    pub storage: PostgresRepo,
     pub flows: HashMap<Uuid, JoinHandle<()>>,
     pub cancellation_token: CancellationToken,
 }
@@ -35,7 +30,7 @@ impl FlowProcessor {
             tx_receiver, 
             flow_receiver, 
             flow_sender, 
-            state: FlowProcessorState { storage }, 
+            storage, 
             flows: HashMap::new(),
             cancellation_token
         }
@@ -70,7 +65,7 @@ impl FlowProcessor {
                             let flow_id = uuid::Uuid::new_v4();
                             
                             let router = FlowProcessorRouter::new(
-                                self.state.clone(), 
+                                self.storage.clone(), 
                                 flow_id,
                                 response_sender,
                                 self.flow_sender.clone()
@@ -110,89 +105,5 @@ impl FlowProcessor {
                 }
             }
         }
-    }
-}
-
-struct FlowProcessorRouter {
-    state: FlowProcessorState,
-    flow_id: Uuid,
-    response_sender: OneshotFlowProcessorSender,
-    task_sender: mpsc::Sender<Uuid>,
-}
-
-impl FlowProcessorRouter {
-    fn new(
-        state: FlowProcessorState,
-        flow_id: Uuid,
-        response_sender: OneshotFlowProcessorSender,
-        task_sender: mpsc::Sender<Uuid>
-    ) -> Self {
-        Self { state, flow_id, response_sender, task_sender }
-    }
-
-    async fn run(mut self, message: FlowProcessorMessage) {
-        let response = match message {
-            FlowProcessorMessage::RunDkgFlow(request) => {
-                let response = self.run_dkg_flow(request).await;
-                let answer = response.map(|response| FlowProcessorResponse::RunDkgFlow(response));
-                answer
-            }
-            FlowProcessorMessage::BridgeRunes(request) => {
-                let response = self.run_bridge_runes_flow(request).await;
-                let answer = response.map(|response| FlowProcessorResponse::BridgeRunes(response));
-                answer
-            }
-            FlowProcessorMessage::ExitSpark(request) => {
-                let response = self.run_exit_spark_flow(request).await;
-                let answer = response.map(|response| FlowProcessorResponse::ExitSpark(response));
-                answer
-            }
-            FlowProcessorMessage::Testing(request) => {
-                let response = self.run_testing_flow(request).await;
-                let answer = response.map(|response| FlowProcessorResponse::Testing(response));
-                answer
-            }
-        };
-
-        let _ = self.response_sender.send(response).map_err(|_| {
-            tracing::error!("[router] Failed to send response for flow id {}", self.flow_id);
-        });
-
-        let _ = self.task_sender.send(self.flow_id).await.map_err(|_| {
-            tracing::error!("[router] Failed to send task for flow id {}", self.flow_id);
-        });
-    }
-
-    async fn run_dkg_flow(&mut self, request: DkgFlowRequest) -> Result<DkgFlowResponse, FlowProcessorError> {
-        Ok(DkgFlowResponse {
-            public_key: format!("[router] public_key for {}", request.request_id),
-        })
-    }
-
-    async fn run_bridge_runes_flow(&mut self, request: BridgeRunesRequest) -> Result<BridgeRunesResponse, FlowProcessorError> {
-        Ok(BridgeRunesResponse {
-            message: format!("message for {}", request.request_id),
-        })
-    }
-
-    async fn run_exit_spark_flow(&mut self, request: ExitSparkRequest) -> Result<ExitSparkResponse, FlowProcessorError> {
-        Ok(ExitSparkResponse {
-            message: format!("message for {}", request.request_id),
-        })
-    }
-
-    async fn run_testing_flow(&mut self, request: TestingRequest) -> Result<TestingResponse, FlowProcessorError> {
-        let mut cur_time = 0;
-        tracing::info!("[router] [{}] Testing flow running for {} seconds", self.flow_id, cur_time);
-        for _ in 0..request.n_runs {
-            tokio::time::sleep(Duration::from_secs(request.n_seconds)).await;
-            cur_time += request.n_seconds;
-            tracing::info!("[router] [{}] Testing flow running for {} seconds", self.flow_id, cur_time);
-        }
-
-        tracing::info!("[router] [{}] Testing flow finished", self.flow_id);
-        Ok(TestingResponse {
-            message: request.message,
-        })
     }
 }
